Restructure tests back to per-topic files
The previous sweep collapsed every topic into a single test/lib/shared.ts,
losing the per-file granularity that the original master/test/*.ts layout
offered. This commit decomposes shared.ts into topic-specific modules under
test/lib/ (sed, chunked, regexp, method, content-type, gzip, synopsis), each
exporting a single run<Topic>Tests(express) function. The two entry points
(sed.express4.test.ts / sed.express5.test.ts) import all seven and run them
as describe blocks. Common helpers (ExpressFn type, document root, sample
file lists, chunkedStatic / concatParser) live in test/lib/util.ts.

Also runs fixpack against package.json (no changes).
