+++ test/sed.express5.test.ts
@@ -1,5 +1,25 @@
 // Test entry for the Express 5 line.
-import express from "express5";
-import {sharedTests} from "./lib/shared.ts";
+// Imports each topic module from test/lib/ and runs them all under a single
+// top-level describe so the source-of-truth granularity matches the original
+// per-topic file layout.
 
-sharedTests(express, "sed.express5.test.ts");
+import {describe} from "node:test";
+import express from "express5";
+
+import {runSedTests} from "./lib/sed.ts";
+import {runChunkedTests} from "./lib/chunked.ts";
+import {runRegexpTests} from "./lib/regexp.ts";
+import {runMethodTests} from "./lib/method.ts";
+import {runContentTypeTests} from "./lib/content-type.ts";
+import {runGzipTests} from "./lib/gzip.ts";
+import {runSynopsisTests} from "./lib/synopsis.ts";
+
+describe("sed.express5.test.ts", () => {
+    runSedTests(express);
+    runChunkedTests(express);
+    runRegexpTests(express);
+    runMethodTests(express);
+    runContentTypeTests(express);
+    runGzipTests(express);
+    runSynopsisTests(express);
+});
